backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
langchain
langchain-community
langchain-openai
chromadb
beautifulsoup4
requests
python-dotenv
tiktoken
pydantic-settings
langchain-chroma
langchain-huggingface
sentence-transformers
google-generativeai
langchain-google-genai
anthropic
langchain-anthropic
langgraph
playwright
greenlet

# AI Agent & Testing
pytest-asyncio
langsmith
ruff
mypy
python-multipart
